+++ Tunny/Settings/Result.cs
@@ -4,5 +4,6 @@ namespace Tunny.Settings
     {
         public string OutputNumberString { get; set; } = "0";
         public int SelectVisualizeType { get; set; } = 3;
+        public int NumberOfClusters { get; set; } = 3;
     }
 }